+++ packages/@lou.codes/constants/src/EMPTY_STRING.ts
@@ -1,0 +1,4 @@
+/**
+ * Empty string.
+ */
+export const EMPTY_STRING = "" as const;