+++ packages/@lou.codes/constants/src/index.ts
@@ -1,0 +1,4 @@
+export { EMPTY_ARRAY } from "./EMPTY_ARRAY.js";
+export { EMPTY_OBJECT } from "./EMPTY_OBJECT.js";
+export { EMPTY_STRING } from "./EMPTY_STRING.js";
+export { createObject } from "./createObject.js";