+++ packages/@lou.codes/create-package/src/validations/hasSpecialCharacters.ts
@@ -1,6 +1,7 @@
 import { foregroundRed } from "@lou.codes/ansi";
+import { EMPTY_STRING } from "@lou.codes/constants";
 
 export const hasSpecialCharacters = (value: string) =>
-	/[~'!()*]/u.test(value.split("/").slice(-1)[0] ?? "") ?
+	/[~'!()*]/u.test(value.split("/").slice(-1)[0] ?? EMPTY_STRING) ?
 		foregroundRed`"${value}" must not contain special characters`
-	:	"";
+	:	EMPTY_STRING;