+++ packages/@lou.codes/create-package/src/validations/hasValidLength.ts
@@ -1,7 +1,8 @@
 import { foregroundRed } from "@lou.codes/ansi";
+import { EMPTY_STRING } from "@lou.codes/constants";
 
 export const hasValidLength =
 	(minimum: number) => (maximum: number) => (value: string) =>
 		value.length >= minimum && value.length <= maximum ?
-			""
+			EMPTY_STRING
 		:	foregroundRed`"${value}" must be between ${minimum} and ${maximum} characters long`;