+++ packages/@lou.codes/iterables/src/iterableToArray.ts
@@ -1,4 +1,9 @@
-import type { IsomorphicIterable, ReadOnlyArray } from "@lou.codes/types";
+import { EMPTY_ARRAY } from "@lou.codes/constants";
+import type {
+	EmptyArray,
+	IsomorphicIterable,
+	ReadOnlyArray,
+} from "@lou.codes/types";
 import { reduce } from "./reduce.js";
 import type { ReadOnlyAsyncIterable } from "./types/ReadOnlyAsyncIterable.js";
 
@@ -13,10 +18,9 @@ import type { ReadOnlyAsyncIterable } from "./types/ReadOnlyAsyncIterable.js";
  * @param iterable Iterable to be turned into an array.
  * @returns Array made of iterable items.
  */
-export const iterableToArray = reduce(item => (array: ReadOnlyArray) => [
-	...array,
-	item,
-])([]) as <Iterable extends IsomorphicIterable>(
+export const iterableToArray = reduce<unknown, ReadOnlyArray | EmptyArray>(
+	item => (array: ReadOnlyArray) => [...array, item],
+)(EMPTY_ARRAY) as <Iterable extends IsomorphicIterable>(
 	iterable: Iterable,
 ) => Iterable extends IsomorphicIterable<infer Item> ?
 	Iterable extends ReadOnlyAsyncIterable<Item> ?