+++ packages/window-open-promise/src/featureParser.ts
@@ -1,4 +1,6 @@
+import { EMPTY_STRING } from "@lou.codes/constants";
 import { objectToEntries } from "@lou.codes/iterables";
+import { always, whenIsUndefined } from "@lou.codes/utils";
 import type { WindowOpenPromiseFeatures } from "./WindowOpenPromiseFeatures.js";
 import { featureJoin } from "./featureJoin.js";
 import { featureMap } from "./featureMap.js";
@@ -17,9 +19,7 @@ import { featureMap } from "./featureMap.js";
  * @param features Features object.
  * @returns Parsed string.
  */
-export const featureParser = (features: WindowOpenPromiseFeatures = {}) =>
-	featureJoin(
-		featureMap(
-			objectToEntries(features as Required<WindowOpenPromiseFeatures>),
-		),
-	);
+export const featureParser = whenIsUndefined(always(EMPTY_STRING))(
+	(features: Required<WindowOpenPromiseFeatures>) =>
+		featureJoin(featureMap(objectToEntries(features))),
+) as (features?: WindowOpenPromiseFeatures) => string;