+++ packages/window-open-promise/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change log
 
+## 1.1.1
+
+### Patch Changes
+
+-   ➕ add @lou.codes/constants and use it.
+-   ⬆️ upgrade `@lou.codes/iterables` to `v1.0.3`
+-   ⬆️ upgrade `@lou.codes/types` to `v1.0.2`
+-   ⬆️ upgrade `@lou.codes/predicates` to `v1.0.2`
+-   ⬆️ upgrade `@lou.codes/utils` to `v1.0.2`
+
 ## 1.1.0
 
 ### Minor Changes